+++ src/test/resources/active_queries_on_sql_server.sql
@@ -28,4 +28,20 @@ SELECT sqltext.TEXT,
 select  session_id, sql.text
 from sys.dm_exec_requests req
        cross apply sys.dm_exec_sql_text(req.sql_handle) sql
-where session_id in (87, 90)
+where session_id in (87, 90);
+
+
+-- Check collation of columns
+SELECT
+  concat('alter table [',  TABLE_SCHEMA, '].[', TABLE_NAME,
+         '] alter column [', COLUMN_NAME, '] ', DATA_TYPE,
+         iif(CHARACTER_MAXIMUM_LENGTH = -1, '(max)', concat('(', CHARACTER_MAXIMUM_LENGTH, ')')),
+         ' COLLATE Latin1_General_CS_AS', iif(IS_NULLABLE = 'NO', ' NOT NULL', ''), ';') correct_collation,
+  TABLE_SCHEMA + '.' + TABLE_NAME,
+  COLUMN_NAME,
+  DATA_TYPE,
+  CHARACTER_MAXIMUM_LENGTH,
+  COLLATION_NAME,
+  IS_NULLABLE
+FROM INFORMATION_SCHEMA.COLUMNS
+where COLLATION_NAME = 'Latin1_General_CI_AS';