+++ CHANGELOG.md
@@ -88,10 +88,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.6.0] (Planned)
 ### Mirror tables
 
+## [1.5.25] - 2023-06-01
+### Added
+- `randstr` ESQL function to generate a random string of arbitrary size, implemented
+  in PostgreSql as `_core.randomstr`.
+- `obfuscate` and `unobfuscate` ESQL function implemented in Postgresql.
+- `checkdigit` ESQL function takes a number and appends a check digit to it computed 
+  using the GTIN-13 method (https://www.gs1.org/services/how-calculate-check-digit-manually).
+- `cast` and `trycast` on dates on SQL Server uses `try_parse` to attempt parsing
+  the date through different cultures until one is found or the cast fails (with
+  an exception in case of `cast` and `null` in case of `trycast`).
+
 ## [1.5.24] - 2023-05-31
 ### Added
 - `randstr` ESQL function to generate a random string of arbitrary size, implemented
-  in SQL Server as `_core.randstr`.
+  in SQL Server as `_core.randomstr`.
 
 ## [1.5.23] - 2023-05-29
 ### Added